Move http server code into it's own module
