+++ tests/components/http/test_cors.py
@@ -31,7 +31,7 @@ TRUSTED_ORIGIN = "https://home-assistant.io"
 
 async def test_cors_middleware_loaded_by_default(hass: HomeAssistant) -> None:
     """Test accessing to server from banned IP when feature is off."""
-    with patch("homeassistant.components.http.setup_cors") as mock_setup:
+    with patch("homeassistant.components.http.server.setup_cors") as mock_setup:
         await async_setup_component(hass, DOMAIN, {"http": {}})
 
     assert len(mock_setup.mock_calls) == 1
@@ -39,7 +39,7 @@ async def test_cors_middleware_loaded_by_default(hass: HomeAssistant) -> None:
 
 async def test_cors_middleware_loaded_from_config(hass: HomeAssistant) -> None:
     """Test accessing to server from banned IP when feature is off."""
-    with patch("homeassistant.components.http.setup_cors") as mock_setup:
+    with patch("homeassistant.components.http.server.setup_cors") as mock_setup:
         await async_setup_component(
             hass,
             DOMAIN,
